requirements.txt
apache-beam[gcp]
pyspark
delta-spark
google-cloud-pubsub
apache-airflow
google-cloud-bigquery
